# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-fam-tree API",
    "description": "Family platform API",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.18"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
...
    "/api/v1/auth/consume": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "consume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Magic link consumed, session established",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Magic link invalid or expired"
          }
        }
      }
    },
...
    "/api/v1/auth/magic-link": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "magic_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Magic link sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkResponseBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid email"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current session claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "No session"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Session refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalid or expired"
          }
        }
      }
    },
...
    "/api/v1/invites/accept": {
      "post": {
        "tags": [
          "invites"
        ],
        "operationId": "accept",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invite accepted, membership inserted, session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Invite token invalid"
          },
          "410": {
            "description": "Invite expired"
          },
          "422": {
            "description": "Existing session belongs to a different email"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptReq": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AcceptRes": {
        "type": "object",
        "required": [
          "family",
          "claims"
        ],
        "properties": {
          "claims": {
            "$ref": "#/components/schemas/ConsumeRes"
          },
          "family": {
            "$ref": "#/components/schemas/FamilyView"
          }
        }
      },
      "AcceptResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AcceptRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "ConsumeReq": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ConsumeRes": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "locale",
          "families"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "families": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyClaim"
            }
          },
          "locale": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ConsumeResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConsumeRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "FamilyClaim": {
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "FamilyView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "properties": {
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Family creation time (RFC 3339). Present on `GET /families/me` so the\nswitcher can disambiguate same-named families by date; `None` on the\nJWT-projection paths (rename, invite-accept) that don't carry it."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "MagicLinkReq": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "Locale the FE detected on the sign-in page (`\"en\"` / `\"de\"`). Optional\nfor backwards-compatible callers; used to seed a brand-new account's\npreference AND to locale-prefix the magic-link URL. For a RETURNING\nuser their stored account locale wins (it's their explicit setting),\nso this only takes effect when the account doesn't exist yet."
          }
        },
        "additionalProperties": false
      },
      "MagicLinkRes": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "MagicLinkResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MagicLinkRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "MeResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConsumeRes"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "description": "Cursor-based pagination block. `next_cursor` is `None` on the last page.",
        "required": [
          "limit",
          "returned"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "returned": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "description": "Per-response metadata. Currently carries pagination cursors, the\n`request_id` for tracing-friendly client diagnostics, and any soft\nvalidation `warnings` attached by the handler.",
        "properties": {
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "description": "Soft warnings. Serialized only when non-empty so the wire shape\nstays unchanged for handlers that don't produce any."
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "admin",
          "owner"
        ]
      },
...
      "Warning": {
        "type": "object",
        "description": "Soft-validation warning.\n\nSurfaced under `meta.warnings` on the success envelope when the\ncaller's input violates a heuristic rule that we don't want to reject\noutright (e.g. parent-child age gap < 14 years). The `code` is a\nstable i18n key (`warning.<snake_case>`); `path` is an optional JSON\nPointer to the relevant field.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "cookie_access": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    }
  },
...
}